Fix reading large sparse matrices
utils.c read_sparse function represents the values of m, n (which are the
dimensions of the matrix) as ints. This causes undefined behaviour when the
value of m*n is greater than the maximum value of an int. This occurs when
reading particularly large sparse matrices. This has been resolved by
promoting the type to "long unsigned".

+++ utils.c
@@ -196,7 +196,7 @@ void read_sparse(const char *file, COO *sparse)
         fclose(f);
         exit(1);
     }
-    if (NZ > m*n) {
+    if (NZ > (long unsigned)m*n) {
         fprintf(stderr, "More nonzeros (%d) than matrix entries (%d x %d)!\n", NZ, m, n);
         fclose(f);
         exit(1);
